Rename a file. Doxygen doesn't like files with the same name.
