+++ libs/Einsums/LinearAlgebra/include/Einsums/LinearAlgebra/GPULinearAlgebra.hpp
@@ -3,7 +3,7 @@
 #include <Einsums/Config.hpp>
 
 #include <Einsums/Concepts/Complex.hpp>
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/Errors/Error.hpp>
 #include <Einsums/GPUStreams/GPUStreams.hpp>
 #include <Einsums/Tensor/DeviceTensor.hpp>
