+++ libs/Einsums/Tensor/include/Einsums/Tensor/DeviceTensor.hpp
@@ -4,7 +4,7 @@
 
 #include <Einsums/Config.hpp>
 
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/DesignPatterns/Lockable.hpp>
 #include <Einsums/Errors/Error.hpp>
 #include <Einsums/Errors/ThrowException.hpp>
