+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/Backends/BlockTileAlgebra.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/TensorAlgebra/Backends/Dispatch.hpp>
 #include <Einsums/TensorAlgebra/Backends/TileAlgebra.hpp>
 #include <Einsums/TensorAlgebra/TensorAlgebra.hpp>
