+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/Backends/DevicePermute.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/Errors/Error.hpp>
 #include <Einsums/GPUStreams/GPUStreams.hpp>
 #include <Einsums/LinearAlgebra.hpp>
