+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/Backends/GenericAlgorithm.hpp
@@ -7,7 +7,7 @@
 
 #include <Einsums/Config.hpp>
 
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/Logging.hpp>
 #include <Einsums/Profile/LabeledSection.hpp>
 #include <Einsums/TensorAlgebra/Detail/Utilities.hpp>
