+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/Backends/KhatriRao.hpp
@@ -7,7 +7,7 @@
 
 #ifndef DOXYGEN
 
-#    include <Einsums/Concepts/Tensor.hpp>
+#    include <Einsums/Concepts/TensorConcepts.hpp>
 #    include <Einsums/Profile/LabeledSection.hpp>
 #    include <Einsums/Tensor/Tensor.hpp>
 #    include <Einsums/TensorAlgebra/Detail/Utilities.hpp>
