+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/Detail/Utilities.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/TensorBase/Common.hpp>
 
 #include <tuple>
