+++ libs/Einsums/TensorAlgebra/include/Einsums/TensorAlgebra/TensorAlgebra.hpp
@@ -9,7 +9,7 @@
 
 #include <Einsums/Concepts/Complex.hpp>
 #include <Einsums/Concepts/SmartPointer.hpp>
-#include <Einsums/Concepts/Tensor.hpp>
+#include <Einsums/Concepts/TensorConcepts.hpp>
 #include <Einsums/Tensor/Tensor.hpp>
 
 #include <cmath>
