+++ libs/Einsums/TensorUtilities/include/Einsums/TensorUtilities/RMSD.hpp
@@ -6,7 +6,7 @@
 #pragma once
 
 #include <Einsums/Tensor/Tensor.hpp>
-#include "Einsums/Concepts/Tensor.hpp"
+#include "Einsums/Concepts/TensorConcepts.hpp"
 
 namespace einsums {
